+++ src/main/java/pingis/controllers/TaskController.java
@@ -65,9 +65,9 @@ public class TaskController {
       return "error";
     }
     
-    if (!taskInstanceService.canContinue(taskInstance, userService.getCurrentUser())) {
+    /*if (!taskInstanceService.canContinue(taskInstance, userService.getCurrentUser())) {
       return "redirect:/error";
-    }
+    }*/
     Challenge currentChallenge = taskInstance.getTask().getChallenge();
     model.addAttribute("challenge", currentChallenge);
     model.addAttribute("task", taskInstance.getTask());
