+++ src/main/java/pingis/services/ChallengeService.java
@@ -67,10 +67,7 @@ public class ChallengeService {
     return challengeRepository.findByName(name);
   }
 
-  public int getNumberOfTasks(Challenge challenge) {
-    List<Task> tasks = taskService.findAllByChallenge(challenge);
-    return tasks.size();
-  }
+
 
   public TaskInstance getUnfinishedTaskInstance(Challenge challenge) {
     User user = userService.getCurrentUser();
@@ -83,64 +80,9 @@ public class ChallengeService {
     return null;
   }
 
-  public boolean isTestTurnInLiveChallenge(Challenge challenge) {
-    User user = userService.getCurrentUser();
-    int side = user.getId() == challenge.getAuthor().getId() ? 0 : 1;
-    int numberOfTasks = getNumberOfTasks(challenge);
-    int highestIndex = numberOfTasks / 2;
-    int numberOfDoneInstances =
-        taskInstanceService.getNumberOfDoneTaskInstancesInChallenge(challenge);
-    if (numberOfDoneInstances % 2 == 0 && highestIndex % 2 == side) {
-      return true;
-    }
-    return false;
-  }
-  
-  public boolean isImplementationTurnInLiveChallenge(Challenge challenge) {
-    User user = userService.getCurrentUser();
-    int side = user.getId() == challenge.getAuthor().getId() ? 0 : 1;
-    int numberOfTasks = getNumberOfTasks(challenge);
-    int highestIndex = numberOfTasks / 2;
-    int numberOfDoneInstances =
-        taskInstanceService.getNumberOfDoneTaskInstancesInChallenge(challenge);
-    if (numberOfDoneInstances % 2 == 1 && highestIndex % 2 == side) {
-      return true;
-    }
-    return false;
-  }
 
-  public boolean isParticipating(Challenge challenge) {
-    User currentUser = userService.getCurrentUser();
-    if (currentUser.getId() == challenge.getAuthor().getId()) {
-      return true;
-    }
 
-    if (challenge.getSecondPlayer() == null) {
-      return false;
-    }
 
-    return (currentUser.getId() == challenge.getSecondPlayer().getId());
-  }
-
-  public Task getTopmostImplementationTask(Challenge challenge, int index) {
-    List<Task> tasks = taskService.findAllByChallenge(challenge);
-    for (Task task : tasks) {
-      if (task.getIndex() == index && task.getType() == TaskType.IMPLEMENTATION) {
-        return task;
-      }
-    }
-    return null;
-  }
-
-  public Task getTopmostTestTask(Challenge challenge, int index) {
-    List<Task> tasks = taskService.findAllByChallenge(challenge);
-    for (Task task : tasks) {
-      if (task.getIndex() == index && task.getType() == TaskType.TEST) {
-        return task;
-      }
-    }
-    return null;
-  }
   
   public Challenge getRandomLiveChallenge(User user) {
     List<Challenge> liveChallenges = findAll().stream()
@@ -158,18 +100,6 @@ public class ChallengeService {
     return liveChallenges.get(0);
   }
 
-  public Challenge getParticipatingLiveChallenge() {
-    Optional<Challenge> findFirst = findAll().stream()
-            .filter(e -> e.getIsOpen())
-            .filter(e -> isParticipating(e))
-            .findFirst();
-    
-    if (findFirst.isPresent()) {
-      return findFirst.get();
-    } else {
-      return null;
-    }
-  }
 
 
 }
