+++ src/main/resources/templates/user.html
@@ -23,7 +23,7 @@
                               <canvas id="progresscircle" width="110" height="110"></canvas>
                         </div>
                   </div>
-
+                  <div th:if="${message != null}" th:text="${message}"></div>
                   <div class="challenge_list row">
                         <div class="col-md-6">
                               <div class="row lead">Started challenges</div>
